lcb: Refactored into separate classes
